Added Cat Sprite and Added Directionality
Replaced generic player sprite with one of the cat assets we found. Currently the cat rotates to face the direction of the player mouse (LOL).

Currently, the cat will display a walk animation when the player is moving with WASD. This is handled independently of the view direction. I tied the player animation speed to a player variable called player.frameLength. This variable decides how frequently a new sprite should be displayed. I am unsure of how frequently the update function increments this value, or rather what dictates when it should increase. I suspect it's not restricted by anything other than processing power, but I haven't researched enough to know for sure.